requirements.txt
absl-py==2.2.2
altair==5.4.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
narwhals==1.38.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.3
scipy==1.10.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.13.3
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.30.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.6
wrapt==1.14.1
zipp==3.20.2
